disable objectMode on output of tap_producer
objectMode is set to false while writableObjectMode is set to true. The
result is that, only when tap_producer act as a Writable, i.e., another
stream is writing into tap_producer, objectMode is used. When
tap_producer is writing string to a consumer, such as process.stdout,
string mode is used.